fix: use CompileImplementation to generate .cmi from .ml files
When a module only has a .ml file (no .mli), we need to compile it
to generate the .cmi file. We were incorrectly using CompileInterface
with a .ml source, which doesn't work.

Now we correctly use CompileImplementation to generate the .cmi from
the .ml file, then compile again to generate the .cmo.